Extract third-party service validation routes into blueprints/validation_routes.py
Phase 2 of breaking up the quickstart.py monolith. Moved the 18 routes
that validate connectivity/credentials for a single third-party
service (Plex, Trakt, MAL, *arr apps, notification services, etc.)
into a new blueprints/ package, registered via Flask's Blueprint
mechanism.

Left in quickstart.py (different concerns, future phases):
- path_validation_rules, validate_library_service_overrides: tied to
  the core library/attribute dependency engine.
- The TMDb/IMDb template-string lookup cluster
  (_get_active_tmdb_api_key, _lookup_tmdb_*, lookup_template_string_value):
  a distinct "resolve a display label" feature, not credential checking.
- validate_all_services: a 590-line orchestrator deeply coupled to the
  core wizard's dependency/validation-summary engine — not a thin
  wrapper, and explicitly scoped for the final, highest-risk phase per
  the refactor plan.

Two small shared utilities the moved routes depended on
(_resolve_request_config_name, utc_now_iso) had no quickstart.py-specific
state, so they moved to modules/persistence.py and modules/helpers.py
respectively, with quickstart.py re-importing them under their original
names — both have other call sites elsewhere in quickstart.py untouched
by this phase, and neither is monkeypatched by tests.

refresh_plex_libraries is also called directly (not just as a route) from
the /step handler, so it's re-imported into quickstart.py's namespace
rather than only registered on the blueprint.

Net: quickstart.py shrinks by 461 lines.
